refactor(table-calc): replace Improve popover with inline AiSlot
The Popover-based Improve UI floated as a disconnected bubble below
the editor with its own rounded border competing with the editor's.
Switch to an inline slot rendered inside the editor container using
the AiSlot presentation that's already used on the SQL side — one
surface at a time, attached to the formula being improved. Adds a
subtle slide-down transition on open and fade on close.
